:sparkles: Added homepage
Set up a few things.
Slim and bootstrap, mostly for comfort.
Set ups are kind of a drag, but I rather this not to look dreadful heh.
No biggie.

Home has a form, that doesn't work as of yet. This is temporary while I
work the details on how to consume the API.